Refactore 1.01
